+++ demisto_sdk/commands/common/update_id_set.py
@@ -112,7 +112,7 @@ def build_tasks_graph(playbook_data):
                 print_warning(f'{playbook_data.get("id")}: No such task {leaf} in playbook')
                 continue
 
-            leaf_next_tasks = sum(leaf_task.get('nexttasks', {}).values(), [])
+            leaf_next_tasks = sum(leaf_task.get('nexttasks', {}).values(), [])  # type: ignore
 
             for task_id in leaf_next_tasks:
                 task = tasks.get(task_id)
@@ -490,12 +490,13 @@ def get_values_for_keys_recursively(json_object: dict, keys_to_search: list) -> 
                     get_values(item)
             return
 
-        for key, value in current_object.items():
-            if isinstance(value, (dict, list)):
-                get_values(value)
-            elif key in keys_to_search:
-                if isinstance(value, (str, int, float, bool)):
-                    values[key].append(value)
+        if isinstance(current_object, dict):
+            for key, value in current_object.items():
+                if isinstance(value, (dict, list)):
+                    get_values(value)
+                elif key in keys_to_search:
+                    if isinstance(value, (str, int, float, bool)):
+                        values[key].append(value)
 
     get_values(json_object)
     return values
@@ -566,7 +567,7 @@ def get_incident_field_data(path, incidents_types_list):
     fromversion = json_data.get('fromVersion')
     toversion = json_data.get('toVersion')
     pack = get_pack_name(path)
-    all_associated_types = set()
+    all_associated_types: set = set()
     all_scripts = set()
 
     associated_types = json_data.get('associatedTypes')
@@ -578,7 +579,7 @@ def get_incident_field_data(path, incidents_types_list):
         all_associated_types = all_associated_types.union(set(system_associated_types))
 
     if 'all' in all_associated_types:
-        all_associated_types = [list(incident_type.keys())[0] for incident_type in incidents_types_list]
+        all_associated_types = {list(incident_type.keys())[0] for incident_type in incidents_types_list}
 
     scripts = json_data.get('script')
     if scripts:
@@ -607,7 +608,7 @@ def get_indicator_type_data(path, all_integrations):
     toversion = json_data.get('toVersion')
     reputation_command = json_data.get('reputationCommand')
     pack = get_pack_name(path)
-    all_scripts = set()
+    all_scripts: set = set()
     associated_integrations = set()
 
     for field in ['reputationScriptName', 'enhancementScriptNames']:
@@ -701,7 +702,7 @@ def get_mapper_data(path):
     toversion = json_data.get('toVersion')
     pack = get_pack_name(path)
     incidents_types = set()
-    incidents_fields = set()
+    incidents_fields: set = set()
 
     default_incident_type = json_data.get('defaultIncidentType')
     if default_incident_type and default_incident_type != '':
@@ -1042,7 +1043,7 @@ class IDSetType(Enum):
 
     @classmethod
     def has_value(cls, value):
-        return value in cls._value2member_map_
+        return value in cls._value2member_map_  # type: ignore
 
 
 class IDSet: