+++ demisto_sdk/commands/lint/linter.py
@@ -77,6 +77,7 @@ class Linter:
             "additional_requirements": [],
             "docker_engine": docker_engine,
             "is_script": False,
+            "commands": None
         }
         # Pack lint status object - visualize it
         self._pkg_lint_status: Dict = {
@@ -192,6 +193,7 @@ class Linter:
                 script_obj = yml_obj.get('script', {}) if isinstance(yml_obj.get('script'), dict) else yml_obj
             self._facts['is_script'] = True if 'Scripts' in yml_file.parts else False
             self._facts['is_long_running'] = script_obj.get('longRunning')
+            self._facts['commands'] = self._get_commands_list(script_obj)
             self._pkg_lint_status["pack_type"] = script_obj.get('type')
         except (FileNotFoundError, IOError, KeyError):
             self._pkg_lint_status["errors"].append('Unable to parse package yml')
@@ -391,6 +393,10 @@ class Linter:
             if py_num < 3:
                 myenv['PY2'] = 'True'
             myenv['is_script'] = str(self._facts['is_script'])
+            # as Xsoar checker is a pylint plugin and runs as part of pylint code, we can not pass args to it.
+            # as a result we can use the env vars as a getway.
+            myenv['commands'] = ','.join([str(elem) for elem in self._facts['commands']]) \
+                if self._facts['commands'] else ''
             stdout, stderr, exit_code = run_command_os(
                 command=build_xsoar_linter_command(lint_files, py_num, self._facts.get('support_level', 'base')),
                 cwd=self._pack_abs_dir, env=myenv)
@@ -736,13 +742,17 @@ class Linter:
         exit_code = SUCCESS
         output = ""
         try:
-            container_obj: docker.models.containers.Container = self._docker_client.containers.run(name=container_name,
-                                                                                                   image=test_image,
-                                                                                                   command=[
-                                                                                                       build_pylint_command(self._facts["lint_files"])],
-                                                                                                   user=f"{os.getuid()}:4000",
-                                                                                                   detach=True,
-                                                                                                   environment=self._facts["env_vars"])
+            container_obj: docker.models.containers.Container = self._docker_client.containers.run(
+                name=container_name,
+                image=test_image,
+                command=[
+                    build_pylint_command(
+                        self._facts["lint_files"], docker_version=self._facts.get('python_version'))
+                ],
+                user=f"{os.getuid()}:4000",
+                detach=True,
+                environment=self._facts["env_vars"]
+            )
             stream_docker_container_output(container_obj.logs(stream=True))
             # wait for container to finish
             container_status = container_obj.wait(condition="exited")
@@ -995,3 +1005,19 @@ class Linter:
             exit_code = RERUN
 
         return exit_code, output
+
+    def _get_commands_list(self, script_obj: dict):
+        """ Get all commands from yml file of the pack
+           Args:
+               script_obj(dict): the script section of the yml file.
+           Returns:
+               list: list of all commands
+        """
+        commands_list = []
+        try:
+            commands_obj = script_obj.get('commands', {})
+            for command in commands_obj:
+                commands_list.append(command.get('name', ''))
+        except Exception:
+            logger.debug("Failed getting the commands from the yml file")
+        return commands_list