+++ demisto_sdk/commands/validate/validate_manager.py
@@ -79,7 +79,7 @@ class ValidateManager:
         self.no_configuration_prints = silence_init_prints
         self.skip_conf_json = skip_conf_json
         self.is_backward_check = is_backward_check
-        self.validate_in_id_set = validate_id_set
+        self.id_set_validations = validate_id_set
         self.is_circle = only_committed_files
         self.validate_all = validate_all
         self.use_git = use_git
@@ -127,7 +127,7 @@ class ValidateManager:
 
         self.id_set_file = self.get_id_set_file(self.skip_id_set_creation, self.id_set_path)
 
-        if self.validate_in_id_set:
+        if self.id_set_validations:
             self.id_set_validator = IDSetValidator(is_circle=self.is_circle, configuration=Configuration())
 
         if no_docker_checks:
@@ -219,7 +219,7 @@ class ValidateManager:
         count = 1
 
         for pack_name in os.listdir(PACKS_DIR):
-            self.completion_percentage = format((count / num_of_packs) * 100, ".2f")
+            self.completion_percentage = format((count / num_of_packs) * 100, ".2f")  # type: ignore
             pack_path = os.path.join(PACKS_DIR, pack_name)
             all_packs_valid.add(self.run_validations_on_pack(pack_path))
             count += 1
@@ -336,9 +336,8 @@ class ValidateManager:
             return True
 
         # id_set validation
-        if self.validate_in_id_set:
-            click.echo(f"Validating id set registration for {file_path}")
-            if not self.id_set_validator.is_file_valid_in_set(file_path):
+        if self.id_set_validations:
+            if not self.id_set_validator.is_file_valid_in_set(file_path, file_type):
                 return False
 
         # Note: these file are not ignored but there are no additional validators for connections
@@ -977,7 +976,7 @@ class ValidateManager:
         old_format_files = set()
         changed_meta_files = set()
         for f in all_files:
-            file_data = list(filter(None, f.split('\t')))
+            file_data: list = list(filter(None, f.split('\t')))
 
             if not file_data:
                 continue
@@ -1112,7 +1111,7 @@ class ValidateManager:
                 self.create_ignored_errors_list(PRESET_ERROR_TO_CHECK.get(key)))
 
     def get_error_ignore_list(self, pack_name):
-        ignored_errors_list = {}
+        ignored_errors_list: dict = {}
         if pack_name:
             pack_ignore_path = get_pack_ignore_file_path(pack_name)
 
@@ -1135,9 +1134,11 @@ class ValidateManager:
         return ignored_errors_list
 
     @staticmethod
-    def get_current_working_branch():
+    def get_current_working_branch() -> str:
         branches = run_command('git branch')
         branch_name_reg = re.search(r'\* (.*)', branches)
+        if not branch_name_reg:
+            return ''
         return branch_name_reg.group(1)
 
     def get_content_release_identifier(self) -> Optional[str]: