+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
+
+
+# 1.2.14
+* SDK repository is now mypy check_untyped_defs complaint.
+* The lint command will now ignore the unsubscriptable-object (E1136) pylint error in dockers based on python 3.9 - this will be removed once a new pylint version is released.
+* Added an option for **format** to run on a whole pack.
+* Added new validation of unimplemented commands from yml in the code to `XSOAR-linter`.
+* Fixed an issue where Auto-Extract fields were only checked for newly added incident types in the **validate** command.
+* Added a new warning validation of direct access to args/params dicts to `XSOAR-linter`.
+
+# 1.2.13
 * Added new validation of indicators usage in CommandResults to `XSOAR-linter`.
 * Running **demisto-sdk lint** will automatically run on changed files (same behavior as the -g flag).
 * Removed supported version message from the documentation when running **generate_docs**.